refactor(analytics): extract context to analyticsContext.ts to fix react-refresh warnings
Moves AnalyticsContextValue, AnalyticsContext and useAnalyticsContext out of
AnalyticsProvider.tsx into a dedicated internal analyticsContext.ts file.
This makes AnalyticsProvider.tsx a pure component file, resolving the two
react-refresh/only-export-components ESLint warnings and improving HMR.

The internal symbols are not re-exported from index.ts, preventing accidental
coupling to @internal APIs by consumers.

Closes #128
