+++ src/analytics/AnalyticsProvider.tsx
@@ -13,21 +13,9 @@
  * ```
  */
 
-import { createContext, useContext, useEffect, useState } from 'react'
+import { useEffect, useState } from 'react'
 import type { ReactNode } from 'react'
-
-/** @internal — brukt av useAnalytics og usePageView */
-export interface AnalyticsContextValue {
-  isEnabled: boolean
-}
-
-/** @internal */
-export const AnalyticsContext = createContext<AnalyticsContextValue>({ isEnabled: false })
-
-/** @internal */
-export function useAnalyticsContext(): AnalyticsContextValue {
-  return useContext(AnalyticsContext)
-}
+import { AnalyticsContext } from './analyticsContext'
 
 /** Props for AnalyticsProvider */
 export interface AnalyticsProviderProps {
